moved uuid expressions to separate file + fix in random
